django_app/requirements.txt
Django==1.8
django-twitter-bootstrap==3.3.0
django-social-auth==0.7.28
